+++ NamedNodeMap.js
@@ -1,23 +1,37 @@
 // import { Node } from './Node.js';
+import { Attr } from './Attr.js';
 
-export class NamedNodeMap {
-	#map;
-
-	constructor(entities) {
-		this.#map = new Map(entities);
+export class NamedNodeMap extends Array {
+	item(n) {
+		return this[n];
 	}
 
 	getNamedItem(name) {
-		return this.#map.get(name);
+		return this.find(attr => attr.name === name);
+	}
+
+	removeNamedItem(name) {
+		const index = this.findIndex(attr => attr.name === name);
+
+		if (index !== -1) {
+			const attr = this.item(index);
+			attr.ownerElement === null;
+			this.splice(index, 1);
+			return attr;
+		}
 	}
 
 	setNamedItem(attr) {
 		if (! (attr instanceof Attr)) {
-			throw new TypeError('Not an Attr.');
+			throw new TypeError('Not an attribute');
+		}
+
+		const index = this.findIndex(attr => attr.name === 'name');
+
+		if (index !== -1) {
+			this[index] =  attr;
 		} else {
-			const oldVal = this.#map.get(attr.name);
-			this.#map.set(attr.name, attr);
-			return oldVal;
+			this.push(attr);
 		}
 	}
 }